Fixed androidTests
